authorization added

+++ src/tasks/controller.py
@@ -17,21 +17,25 @@ def create_task(body:TaskShema,db:Session,user:UserModel):
     db.refresh(new_task)
     return new_task
 
-def get_task(db:Session):
-    tasks=db.query(TaskModel).all()
+def get_task(db:Session,user:UserModel):
+    tasks=db.query(TaskModel).filter(TaskModel.user_id==user.id).all()
     return tasks
 
-def get_one_task(task_id:int,db:Session):
+def get_one_task(task_id:int,db:Session,user:UserModel):
     one_task=db.query(TaskModel).get(task_id)
     if not one_task:
         raise HTTPException(404,detail="Task id not incorrect")
+    if one_task.user_id!=user.id:
+        raise HTTPException(401,detail="you are not authorized")
     return one_task 
 
-def update_task(body:TaskShema,task_id:int,db:Session):
-    one_task=db.query(TaskModel).get(task_id)
+def update_task(body:TaskShema,task_id:int,db:Session,user:UserModel):
+    one_task:TaskModel=db.query(TaskModel).get(task_id)
 
     if not one_task:
         raise HTTPException(404,detail="Task id incorrect")
+    if one_task.user_id!=user.id:
+        raise HTTPException(401,detail="you are not allowed to update this task")
     
     # one_task.title=body.title
     # one_task.description=body.description
@@ -47,10 +51,12 @@ def update_task(body:TaskShema,task_id:int,db:Session):
 
     return one_task
 
-def delete_task(task_id:int,db:Session):
+def delete_task(task_id:int,db:Session,user:UserModel):
     one_task=db.query(TaskModel).get(task_id)
     if not one_task:
         raise HTTPException(404,detail="incorrect task_id")
+    if one_task.user_id!=user.id:
+        raise HTTPException(401,detail="you are not allowed to delete this task")
     
     db.delete(one_task)
     db.commit()
